chore: entity 정리
