+++ KnockKnock-iOS/Scenes/Feed/FeedDetail/Views/FeedDetailNavigtaionBarView.swift
@@ -61,7 +61,7 @@ final class FeedDetailNavigationBarView: UIView {
   
   // MARK: - Bind
   
-  func bind(feed: FeedPost) {
+  func bind(feed: FeedDetail.Post) {
     self.userIdLabel.text = feed.userName
     self.postDateLabel.text = feed.regDate
     if let userImage = feed.userImage {
