+++ KnockKnock-iOS/Scenes/Feed/FeedList/Views/FeedListHeaderReusableView.swift
@@ -77,7 +77,7 @@ class FeedListHeaderReusableView: UICollectionReusableView {
 
   // MARK: - Bind
 
-  func bind(feed: FeedListPost) {
+  func bind(feed: FeedList.Post) {
     self.userIdLabel.text = "\(feed.id)" // feed.userName으로 변경할 것
     self.postDateLabel.text = feed.regDateToString
     self.profileImageView.setImageFromStringUrl(
